remove extra log + rename test
