+++ .github/workflows/unitary_tests.yml
@@ -33,6 +33,6 @@ jobs:
       run: |
         python -m unittest tests/test_hashes.py
         python -m unittest tests/test_routing.py
-        python -m unittest tests/test_dht.py
+        python -m unittest tests/test_network.py
