+++ tests/__init__.py
@@ -1,3 +1,3 @@
 from tests.test_hashes import *
 from tests.test_routing import *
-from tests.test_dht import *
+from tests.test_network import *
